Lesson 20 Add JUnit

+++ ITCources/src/main/java/ru/tsedrik/service/CourseServiceImpl.java
@@ -1,17 +1,13 @@
 package ru.tsedrik.service;
 
-import org.hibernate.SessionFactory;
 import ru.tsedrik.dao.CourseDAO;
 import ru.tsedrik.dao.StudentDAO;
-import ru.tsedrik.dao.jdbc.CourseDAOImpl;
-import ru.tsedrik.dao.jdbc.StudentDAOImpl;
 import ru.tsedrik.entity.Course;
 import ru.tsedrik.entity.CourseType;
 import ru.tsedrik.entity.Student;
 
 import java.util.Collection;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 public class CourseServiceImpl implements CourseService{
 
@@ -19,34 +15,49 @@ public class CourseServiceImpl implements CourseService{
 
     private StudentDAO studentDAO;
 
-    public CourseServiceImpl(SessionFactory sessionFactory) {
-        this.courseDAO = new CourseDAOImpl(sessionFactory);
-        this.studentDAO = new StudentDAOImpl(sessionFactory);
+    public CourseServiceImpl(CourseDAO courseDAO, StudentDAO studentDAO) {
+        this.courseDAO = courseDAO;
+        this.studentDAO = studentDAO;
     }
 
     @Override
     public Course addCourse(Course course) {
-
-       return courseDAO.create(course);
+        if (course == null){
+            throw new IllegalArgumentException("Added course can't be null.");
+        }
+        return courseDAO.create(course);
     }
 
     @Override
     public boolean deleteCourse(Course course) {
+        if (course == null || course.getId() == null){
+            throw new IllegalArgumentException("Deleted course and its id can't be null.");
+        }
         return courseDAO.delete(course);
     }
 
     @Override
     public boolean deleteCourseById(UUID id) {
+        if (id == null){
+            throw new IllegalArgumentException("Deleted course's id can't be null.");
+        }
         return courseDAO.deleteById(id);
     }
 
     @Override
     public Collection<Course> getCourseByType(CourseType type) {
+        if (type == null){
+            throw new IllegalArgumentException("Type can't be null.");
+        }
         return courseDAO.getByCourseType(type);
     }
 
     @Override
     public Course getCourseById(UUID id) {
+        if (id == null){
+            throw new IllegalArgumentException("Course's id can't be null.");
+        }
+
         Course course = courseDAO.getById(id);
 
         if (course == null){
@@ -61,6 +72,11 @@ public class CourseServiceImpl implements CourseService{
     }
 
     public boolean enroll(UUID courseId, String email){
+
+        if (courseId == null || email == null || email.isEmpty()){
+            throw new IllegalArgumentException("Course's id and email can't be null or empty.");
+        }
+
         Course course = courseDAO.getById(courseId);
 
         if (course == null){
