+++ ITCources/src/main/java/ru/tsedrik/service/StudentServiceImpl.java
@@ -1,8 +1,6 @@
 package ru.tsedrik.service;
 
-import org.hibernate.SessionFactory;
 import ru.tsedrik.dao.StudentDAO;
-import ru.tsedrik.dao.jdbc.StudentDAOImpl;
 import ru.tsedrik.entity.Student;
 
 import java.util.UUID;
@@ -11,19 +9,24 @@ public class StudentServiceImpl implements StudentService{
 
     private StudentDAO studentDAO;
 
-    public StudentServiceImpl(SessionFactory sessionFactory){
-        this.studentDAO = new StudentDAOImpl(sessionFactory);
+    public StudentServiceImpl(StudentDAO studentDAO){
+        this.studentDAO = studentDAO;
     }
 
     @Override
     public Student addStudent(String email) {
+        if (email == null){
+            throw new IllegalArgumentException("Email can't be null.");
+        }
         Student student = new Student(UUID.randomUUID(), email);
-        studentDAO.create(student);
-        return student;
+        return studentDAO.create(student);
     }
 
     @Override
     public Student getStudentById(UUID id) {
+        if (id == null){
+            throw new IllegalArgumentException("Student's id can't be null.");
+        }
         Student student = studentDAO.getById(id);
         if (student == null){
             throw new RuntimeException("There is no student with id = " + id);
@@ -33,11 +36,17 @@ public class StudentServiceImpl implements StudentService{
 
     @Override
     public boolean deleteStudentById(UUID id) {
+        if (id == null){
+            throw new IllegalArgumentException("Student's id can't be null.");
+        }
         return studentDAO.deleteById(id);
     }
 
     @Override
     public boolean deleteStudent(Student student) {
+        if (student == null || student.getId() == null){
+            throw new IllegalArgumentException("Deleted student and its id can't be null.");
+        }
         return studentDAO.delete(student);
     }
